+++ src/main/java/uk/co/shockwaveinteractive/client/screen/GuiShieldLevel.java
@@ -4,7 +4,6 @@ import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiComponent;
-import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.client.gui.overlay.ForgeGui;
@@ -33,11 +32,11 @@ public class GuiShieldLevel extends GuiComponent {
             ||  ( !(minecraft.player.getMainHandItem().getItem() instanceof ItemShieldModule) && (shieldModule.getTotalDamageTaken(stack) == 0 && currentTime - shieldModule.getLastDamageTime(stack) > (ItemShieldModule.BASE_SHIELD_RECHARGE_DELAY + 5) * 20)))
             return;
 
-        var allShields = ItemShieldModule.getShieldsInInventory(minecraft.player, false);
-        var activeShields = ItemShieldModule.getShieldsInInventory(minecraft.player, true);
+//        var allShields = ItemShieldModule.getShieldsInInventory(minecraft.player, false);
+//        var activeShields = ItemShieldModule.getShieldsInInventory(minecraft.player, true);
 
-        int maxShieldBuffer = ItemShieldModule.getDamageThreshold(activeShields);
-        int rechargeDelay = ItemShieldModule.getRechargeDelay(activeShields);
+        int maxShieldBuffer = ItemShieldModule.getDamageThreshold();
+        int rechargeDelay = ItemShieldModule.getRechargeDelay();
         int yOffset = 20;
         int offsetX = Math.round((float) minecraft.getWindow().getGuiScaledWidth() / 2) - 54;
         int barWidth = 108;
@@ -56,19 +55,19 @@ public class GuiShieldLevel extends GuiComponent {
             );
 
         // Energy Percentage
-        float energyPercentage = Math.round((ItemShieldModule.getTotalEnergyRemaining(activeShields) / ItemShieldModule.getTotalMaxEnergy(allShields)) * 100);
+        float energyPercentage = Math.round((ItemShieldModule.getEnergyStored(stack) / ItemShieldModule.getTotalMaxEnergy()) * 100);
         String energyPercentageText = energyPercentage + "%";
         int maxenergyWidth = minecraft.font.width(energyPercentageText);
         int offsetEnergy = offsetX - 15 - maxenergyWidth / 2 + (maxenergyWidth - minecraft.font.width(energyPercentageText));
         drawString(ms, minecraft.font, energyPercentageText, offsetEnergy, yOffset - 10, 0xFFFFFF);
 
         // Active Shields
-        if(allShields.size() > 1) {
-            String textActive = Component.translatable("shockmetal.tooltip.energy.active", activeShields.size() + " / " + allShields.size()).getString();
-            int maxWidthActive = minecraft.font.width(maxShieldBuffer + " / " + maxShieldBuffer);
-            int offsetActive = offsetX + 155 - maxWidthActive / 2 + (maxWidthActive - minecraft.font.width(textActive));
-            drawString(ms, minecraft.font, textActive, offsetActive, yOffset - 10, 0xFFFFFF);
-        }
+//        if(allShields.size() > 1) {
+//            String textActive = Component.translatable("shockmetal.tooltip.energy.active", activeShields.size() + " / " + allShields.size()).getString();
+//            int maxWidthActive = minecraft.font.width(maxShieldBuffer + " / " + maxShieldBuffer);
+//            int offsetActive = offsetX + 155 - maxWidthActive / 2 + (maxWidthActive - minecraft.font.width(textActive));
+//            drawString(ms, minecraft.font, textActive, offsetActive, yOffset - 10, 0xFFFFFF);
+//        }
 
         // Energy Buffer
         String text = (maxShieldBuffer - shieldModule.getTotalDamageTaken(stack)) + " / " + maxShieldBuffer;