+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Version 1.19.2-3.0.2.0
 ### Additions
-  * Added new item - Shield Module, can protect the player by absorbing damage 
+ * Added new item - Shield Module, can protect the player by absorbing damage 
 ### Changes
-
+ * Increased Tetra mining level
 ### Fixes